Finish extended Chain DSL API
Chains can now have additional constraints assigned when using the
kotlin DSL.

Useful to constrain the chain to anything other than the parent.
