+++ constraintlayout/compose/src/main/java/androidx/constraintlayout/compose/ConstraintLayout.kt
@@ -20,7 +20,6 @@ import android.annotation.SuppressLint
 import android.os.Handler
 import android.os.Looper
 import android.util.Log
-import androidx.annotation.FloatRange
 import androidx.compose.animation.core.Animatable
 import androidx.compose.animation.core.AnimationSpec
 import androidx.compose.animation.core.tween
@@ -50,7 +49,6 @@ import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.unit.*
 import androidx.compose.ui.util.fastForEach
 import androidx.compose.ui.util.fastForEachIndexed
-import androidx.constraintlayout.core.SolverVariable
 import androidx.constraintlayout.core.parser.CLObject
 import androidx.constraintlayout.core.parser.CLParser
 import androidx.constraintlayout.core.parser.CLParsingException
@@ -431,14 +429,6 @@ class ConstraintSetScope internal constructor() : ConstraintLayoutBaseScope() {
         ref: ConstrainedLayoutReference,
         constrainBlock: ConstrainScope.() -> Unit
     ) = ConstrainScope(ref.id).apply {
-        constrainBlock()
-        this@ConstraintSetScope.tasks.addAll(this.tasks)
-    }
-
-    fun constrain(
-        ref: HorizontalChainReference,
-        constrainBlock: HorizontalChainScope.() -> Unit
-    ) = HorizontalChainScope(ref.id).apply {
         constrainBlock()
         this@ConstraintSetScope.tasks.addAll(this.tasks)
     }
