+++ constraintlayout/compose/src/main/java/androidx/constraintlayout/compose/ConstraintLayoutBaseScope.kt
@@ -16,17 +16,12 @@
 
 package androidx.constraintlayout.compose
 
-import androidx.annotation.FloatRange
-import androidx.compose.foundation.layout.LayoutScopeMarker
 import androidx.compose.runtime.Immutable
 import androidx.compose.runtime.Stable
-import androidx.compose.ui.graphics.TransformOrigin
 import androidx.compose.ui.layout.FirstBaseline
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.LayoutDirection
 import androidx.compose.ui.unit.dp
-import androidx.constraintlayout.core.state.ConstraintReference
-import androidx.constraintlayout.core.state.helpers.HorizontalChainReference
 import androidx.constraintlayout.core.widgets.ConstraintWidget
 
 /**
@@ -44,7 +39,9 @@ abstract class ConstraintLayoutBaseScope {
         helpersHashCode = 0
     }
 
-    @PublishedApi internal var helpersHashCode: Int = 0
+    @PublishedApi
+    internal var helpersHashCode: Int = 0
+
     private fun updateHelpersHashCode(value: Int) {
         helpersHashCode = (helpersHashCode * 1009 + value) % 1000000007
     }
@@ -74,6 +71,29 @@ abstract class ConstraintLayoutBaseScope {
     // TODO(popam): investigate if this can be just a HorizontalAnchor
     @Stable
     data class BaselineAnchor internal constructor(internal val id: Any)
+
+
+    /**
+     * Specifies additional constraints associated to the horizontal chain identified with [ref].
+     */
+    fun constrain(
+        ref: HorizontalChainReference,
+        constrainBlock: HorizontalChainScope.() -> Unit
+    ) = HorizontalChainScope(ref.id).apply {
+        constrainBlock()
+        this@ConstraintLayoutBaseScope.tasks.addAll(this.tasks)
+    }
+
+    /**
+     * Specifies additional constraints associated to the vertical chain identified with [ref].
+     */
+    fun constrain(
+        ref: VerticalChainReference,
+        constrainBlock: VerticalChainScope.() -> Unit
+    ) = VerticalChainScope(ref.id).apply {
+        constrainBlock()
+        this@ConstraintLayoutBaseScope.tasks.addAll(this.tasks)
+    }
 
     /**
      * Creates a guideline at a specific offset from the start of the [ConstraintLayout].
@@ -350,15 +370,21 @@ abstract class ConstraintLayoutBaseScope {
 
     /**
      * Creates a horizontal chain including the referenced layouts.
+     *
+     * Use [constrain] with the resulting [HorizontalChainReference] to modify the start/left and
+     * end/right constraints of this chain.
      */
     // TODO(popam, b/157783937): this API should be improved
     fun createHorizontalChain(
         vararg elements: ConstrainedLayoutReference,
         chainStyle: ChainStyle = ChainStyle.Spread
-    ): androidx.constraintlayout.compose.HorizontalChainReference {
+    ): HorizontalChainReference {
         val id = createHelperId()
         tasks.add { state ->
-            val helper = state.helper(id, androidx.constraintlayout.core.state.State.Helper.HORIZONTAL_CHAIN) as HorizontalChainReference
+            val helper = state.helper(
+                id,
+                androidx.constraintlayout.core.state.State.Helper.HORIZONTAL_CHAIN
+            ) as androidx.constraintlayout.core.state.helpers.HorizontalChainReference
             helper.add(*(elements.map { it.id }.toTypedArray()))
             helper.style(chainStyle.style)
             helper.apply()
@@ -369,21 +395,29 @@ abstract class ConstraintLayoutBaseScope {
         updateHelpersHashCode(16)
         elements.forEach { updateHelpersHashCode(it.hashCode()) }
         updateHelpersHashCode(chainStyle.hashCode())
-        return androidx.constraintlayout.compose.HorizontalChainReference(id)
+        return HorizontalChainReference(id)
     }
 
     /**
      * Creates a vertical chain including the referenced layouts.
+     *
+     * Use [constrain] with the resulting [VerticalChainReference] to modify the top and
+     * bottom constraints of this chain.
      */
     // TODO(popam, b/157783937): this API should be improved
     fun createVerticalChain(
         vararg elements: ConstrainedLayoutReference,
         chainStyle: ChainStyle = ChainStyle.Spread
-    ) {
+    ): VerticalChainReference {
+        val id = createHelperId()
         tasks.add { state ->
-            state.verticalChain(*(elements.map { it.id }.toTypedArray()))
-                .also { it.style(chainStyle.style) }
-                .apply()
+            val helper = state.helper(
+                id,
+                androidx.constraintlayout.core.state.State.Helper.VERTICAL_CHAIN
+            ) as androidx.constraintlayout.core.state.helpers.VerticalChainReference
+            helper.add(*(elements.map { it.id }.toTypedArray()))
+            helper.style(chainStyle.style)
+            helper.apply()
             if (chainStyle.bias != null) {
                 state.constraints(elements[0].id).verticalBias(chainStyle.bias)
             }
@@ -391,6 +425,7 @@ abstract class ConstraintLayoutBaseScope {
         updateHelpersHashCode(17)
         elements.forEach { updateHelpersHashCode(it.hashCode()) }
         updateHelpersHashCode(chainStyle.hashCode())
+        return VerticalChainReference(id)
     }
 }
 
@@ -442,31 +477,60 @@ class ConstrainedLayoutReference(val id: Any) {
     val baseline = ConstraintLayoutBaseScope.BaselineAnchor(id)
 }
 
+/**
+ * Represents a horizontal chain within a [ConstraintLayout].
+ *
+ * The anchors correspond to the first and last elements in the chain.
+ */
 @Stable
 class HorizontalChainReference internal constructor(internal val id: Any) {
     /**
-     * The start anchor of this layout. Represents left in LTR layout direction, or right in RTL.
+     * The start anchor of the first element in the chain.
+     *
+     * Represents left in LTR layout direction, or right in RTL.
      */
     @Stable
     val start = ConstraintLayoutBaseScope.VerticalAnchor(id, -2)
 
     /**
-     * The left anchor of this layout.
+     * The left anchor of the first element in the chain.
      */
     @Stable
     val absoluteLeft = ConstraintLayoutBaseScope.VerticalAnchor(id, 0)
 
     /**
-     * The end anchor of this layout. Represents right in LTR layout direction, or left in RTL.
+     * The end anchor of the last element in the chain.
+     *
+     * Represents right in LTR layout direction, or left in RTL.
      */
     @Stable
     val end = ConstraintLayoutBaseScope.VerticalAnchor(id, -1)
 
     /**
-     * The right anchor of this layout.
+     * The right anchor of the last element in the chain.
      */
     @Stable
     val absoluteRight = ConstraintLayoutBaseScope.VerticalAnchor(id, 1)
+}
+
+/**
+ * Represents a vertical chain within a [ConstraintLayout].
+ *
+ * The anchors correspond to the first and last elements in the chain.
+ */
+@Stable
+class VerticalChainReference internal constructor(internal val id: Any) {
+    /**
+     * The top anchor of the first element in the chain.
+     */
+    @Stable
+    val top = ConstraintLayoutBaseScope.HorizontalAnchor(id, 0)
+
+    /**
+     * The bottom anchor of the last element in the chain.
+     */
+    @Stable
+    val bottom = ConstraintLayoutBaseScope.HorizontalAnchor(id, 1)
 }
 
 /**
@@ -537,621 +601,5 @@ class Visibility internal constructor(
          */
         @Stable
         val Gone = Visibility(ConstraintWidget.GONE)
-    }
-}
-
-/**
- * Scope that can be used to constrain a layout.
- *
- * Used within `Modifier.constrainAs` from the inline DSL API. And within `constrain` from the
- * ConstraintSet-based API.
- */
-@LayoutScopeMarker
-@Stable
-class ConstrainScope internal constructor(internal val id: Any) {
-    internal val tasks = mutableListOf<(State) -> Unit>()
-    internal fun applyTo(state: State) = tasks.forEach { it(state) }
-
-    /**
-     * Reference to the [ConstraintLayout] itself, which can be used to specify constraints
-     * between itself and its children.
-     */
-    val parent = ConstrainedLayoutReference(SolverState.PARENT)
-
-    /**
-     * The start anchor of the layout - can be constrained using [VerticalAnchorable.linkTo].
-     */
-    val start: VerticalAnchorable = ConstraintVerticalAnchorable(id, -2)
-
-    /**
-     * The left anchor of the layout - can be constrained using [VerticalAnchorable.linkTo].
-     */
-    val absoluteLeft: VerticalAnchorable = ConstraintVerticalAnchorable(id, 0)
-
-    /**
-     * The top anchor of the layout - can be constrained using [HorizontalAnchorable.linkTo].
-     */
-    val top: HorizontalAnchorable = ConstraintHorizontalAnchorable(id, 0)
-
-    /**
-     * The end anchor of the layout - can be constrained using [VerticalAnchorable.linkTo].
-     */
-    val end: VerticalAnchorable = ConstraintVerticalAnchorable(id, -1)
-
-    /**
-     * The right anchor of the layout - can be constrained using [VerticalAnchorable.linkTo].
-     */
-    val absoluteRight: VerticalAnchorable = ConstraintVerticalAnchorable(id, 1)
-
-    /**
-     * The bottom anchor of the layout - can be constrained using [HorizontalAnchorable.linkTo].
-     */
-    val bottom: HorizontalAnchorable = ConstraintHorizontalAnchorable(id, 1)
-
-    /**
-     * The [FirstBaseline] of the layout - can be constrained using [BaselineAnchorable.linkTo].
-     */
-    val baseline: BaselineAnchorable = ConstraintBaselineAnchorable(id)
-
-    /**
-     * The width of the [ConstraintLayout] child.
-     */
-    var width: Dimension = Dimension.wrapContent
-        set(value) {
-            field = value
-            tasks.add { state ->
-                state.constraints(id).width(
-                    (value as DimensionDescription).toSolverDimension(state)
-                )
-            }
-        }
-
-    /**
-     * The height of the [ConstraintLayout] child.
-     */
-    var height: Dimension = Dimension.wrapContent
-        set(value) {
-            field = value
-            tasks.add { state ->
-                state.constraints(id).height(
-                    (value as DimensionDescription).toSolverDimension(state)
-                )
-            }
-        }
-
-    /**
-     * The overall visibility of the [ConstraintLayout] child.
-     *
-     * [Visibility.Visible] by default.
-     */
-    var visibility: Visibility = Visibility.Visible
-        set(value) {
-            field = value
-            tasks.add { state ->
-                with(state.constraints(id)) {
-                    visibility(value.solverValue)
-                    if (value == Visibility.Invisible) {
-                        // A bit of a hack, this behavior is not defined in :core
-                        // Invisible should override alpha
-                        alpha(0f)
-                    }
-                }
-            }
-        }
-
-    /**
-     * The transparency value when rendering the content.
-     */
-    @FloatRange(from = 0.0, to = 1.0)
-    var alpha: Float = 1.0f
-        set(value) {
-            field = value
-            addTransform {
-                if (visibility != Visibility.Invisible) {
-                    // A bit of a hack, this behavior is not defined in :core
-                    // Invisible should override alpha
-                    alpha(value)
-                }
-            }
-        }
-
-    /**
-     * The percent scaling value on the horizontal axis. Where 1 is 100%.
-     */
-    var scaleX: Float = 1.0f
-        set(value) {
-            field = value
-            addTransform { scaleX(value) }
-        }
-
-    /**
-     * The percent scaling value on the vertical axis. Where 1 is 100%.
-     */
-    var scaleY: Float = 1.0f
-        set(value) {
-            field = value
-            addTransform { scaleY(value) }
-        }
-
-
-    /**
-     * The degrees to rotate the content over the horizontal axis.
-     */
-    var rotationX: Float = 0.0f
-        set(value) {
-            field = value
-            addTransform { rotationX(value) }
-        }
-
-    /**
-     * The degrees to rotate the content over the vertical axis.
-     */
-    var rotationY: Float = 0.0f
-        set(value) {
-            field = value
-            addTransform { rotationY(value) }
-        }
-
-    /**
-     * The degrees to rotate the content on the screen plane.
-     */
-    var rotationZ: Float = 0.0f
-        set(value) {
-            field = value
-            addTransform { rotationZ(value) }
-        }
-
-    /**
-     * The distance to offset the content over the X axis.
-     */
-    var translationX: Dp = 0.dp
-        set(value) {
-            field = value
-            addFloatTransformFromDp(value) { floatValue -> translationX(floatValue) }
-        }
-
-    /**
-     * The distance to offset the content over the Y axis.
-     */
-    var translationY: Dp = 0.dp
-        set(value) {
-            field = value
-            addFloatTransformFromDp(value) { floatValue -> translationY(floatValue) }
-        }
-
-    /**
-     * The distance to offset the content over the Z axis.
-     */
-    var translationZ: Dp = 0.dp
-        set(value) {
-            field = value
-            addFloatTransformFromDp(value) { floatValue -> translationZ(floatValue) }
-        }
-
-    /**
-     * The X axis offset percent where the content is rotated and scaled.
-     *
-     * @see [TransformOrigin]
-     */
-    var pivotX: Float = 0.5f
-        set(value) {
-            field = value
-            addTransform { pivotX(value) }
-        }
-
-    /**
-     * The Y axis offset percent where the content is rotated and scaled.
-     *
-     * @see [TransformOrigin]
-     */
-    var pivotY: Float = 0.5f
-        set(value) {
-            field = value
-            addTransform { pivotY(value) }
-        }
-
-    /**
-     * Whenever the width is not fixed, this weight may be used by an horizontal Chain to decide how
-     * much space assign to this widget.
-     */
-    var horizontalChainWeight: Float = Float.NaN
-        set(value) {
-            field = value
-            tasks.add { state ->
-                state.constraints(id).horizontalChainWeight = value
-            }
-        }
-
-    /**
-     * Whenever the height is not fixed, this weight may be used by a vertical Chain to decide how
-     * much space assign to this widget.
-     */
-    var verticalChainWeight: Float = Float.NaN
-        set(value) {
-            field = value
-            tasks.add { state ->
-                state.constraints(id).verticalChainWeight = value
-            }
-        }
-
-    private fun addTransform(change: ConstraintReference.() -> Unit) =
-        tasks.add { state -> change(state.constraints(id)) }
-
-    private fun addFloatTransformFromDp(dpValue: Dp, change: ConstraintReference.(Float) -> Unit) =
-        tasks.add { state ->
-            (state as? State)?.also {
-                state.constraints(id).change(state.convertDimension(dpValue).toFloat())
-            }
-        }
-
-
-    /**
-     * Represents a vertical side of a layout (i.e start and end) that can be anchored using
-     * [linkTo] in their `Modifier.constrainAs` blocks.
-     */
-    internal inner class ConstraintVerticalAnchorable internal constructor(
-        internal val id: Any,
-        internal val index: Int
-    ): VerticalAnchorable {
-        /**
-         * Adds a link towards a [ConstraintLayoutBaseScope.VerticalAnchor].
-         */
-        // TODO(popam, b/158069248): add parameter for gone margin
-        override fun linkTo(anchor: ConstraintLayoutBaseScope.VerticalAnchor, margin: Dp) {
-            tasks.add { state ->
-                with(state.constraints(id)) {
-                    val layoutDirection = state.layoutDirection
-                    val index1 = verticalAnchorIndexToFunctionIndex(index, layoutDirection)
-                    val index2 = verticalAnchorIndexToFunctionIndex(anchor.index, layoutDirection)
-                    verticalAnchorFunctions[index1][index2]
-                        .invoke(this, anchor.id, state.layoutDirection)
-                        .margin(margin)
-                }
-            }
-        }
-    }
-
-    /**
-     * Represents a horizontal side of a layout (i.e top and bottom) that can be anchored using
-     * [linkTo] in their `Modifier.constrainAs` blocks.
-     */
-    internal inner class ConstraintHorizontalAnchorable internal constructor(
-        internal val tag: Any,
-        internal val index: Int
-    ): HorizontalAnchorable {
-        /**
-         * Adds a link towards a [ConstraintLayoutBaseScope.HorizontalAnchor].
-         */
-        // TODO(popam, b/158069248): add parameter for gone margin
-        override fun linkTo(anchor: ConstraintLayoutBaseScope.HorizontalAnchor, margin: Dp) {
-            tasks.add { state ->
-                with(state.constraints(id)) {
-                    horizontalAnchorFunctions[index][anchor.index]
-                        .invoke(this, anchor.id)
-                        .margin(margin)
-                }
-            }
-        }
-    }
-
-    /**
-     * Represents the [FirstBaseline] of a layout that can be anchored
-     * using [linkTo] in their `Modifier.constrainAs` blocks.
-     */
-    internal inner class ConstraintBaselineAnchorable internal constructor(
-        internal val id: Any
-        ): BaselineAnchorable {
-        /**
-         * Adds a link towards a [ConstraintLayoutBaseScope.BaselineAnchor].
-         */
-        // TODO(popam, b/158069248): add parameter for gone margin
-        override fun linkTo(anchor: ConstraintLayoutBaseScope.BaselineAnchor, margin: Dp) {
-            tasks.add { state ->
-                (state as? State)?.let {
-                    it.baselineNeededFor(id)
-                    it.baselineNeededFor(anchor.id)
-                }
-                with(state.constraints(id)) {
-                    baselineAnchorFunction.invoke(this, anchor.id).margin(margin)
-                }
-            }
-        }
-    }
-
-    /**
-     * Adds both start and end links towards other [ConstraintLayoutBaseScope.VerticalAnchor]s.
-     */
-    // TODO(popam, b/158069248): add parameter for gone margin
-    fun linkTo(
-        start: ConstraintLayoutBaseScope.VerticalAnchor,
-        end: ConstraintLayoutBaseScope.VerticalAnchor,
-        startMargin: Dp = 0.dp,
-        endMargin: Dp = 0.dp,
-        @FloatRange(from = 0.0, to = 1.0) bias: Float = 0.5f
-    ) {
-        this@ConstrainScope.start.linkTo(start, startMargin)
-        this@ConstrainScope.end.linkTo(end, endMargin)
-        tasks.add { state ->
-            val resolvedBias = if (state.layoutDirection == LayoutDirection.Rtl) 1 - bias else bias
-            state.constraints(id).horizontalBias(resolvedBias)
-        }
-    }
-
-    /**
-     * Adds both top and bottom links towards other [ConstraintLayoutBaseScope.HorizontalAnchor]s.
-     */
-    // TODO(popam, b/158069248): add parameter for gone margin
-    fun linkTo(
-        top: ConstraintLayoutBaseScope.HorizontalAnchor,
-        bottom: ConstraintLayoutBaseScope.HorizontalAnchor,
-        topMargin: Dp = 0.dp,
-        bottomMargin: Dp = 0.dp,
-        @FloatRange(from = 0.0, to = 1.0) bias: Float = 0.5f
-    ) {
-        this@ConstrainScope.top.linkTo(top, topMargin)
-        this@ConstrainScope.bottom.linkTo(bottom, bottomMargin)
-        tasks.add { state ->
-            state.constraints(id).verticalBias(bias)
-        }
-    }
-
-    /**
-     * Adds all start, top, end, bottom links towards
-     * other [ConstraintLayoutBaseScope.HorizontalAnchor]s.
-     */
-    // TODO(popam, b/158069248): add parameter for gone margin
-    fun linkTo(
-        start: ConstraintLayoutBaseScope.VerticalAnchor,
-        top: ConstraintLayoutBaseScope.HorizontalAnchor,
-        end: ConstraintLayoutBaseScope.VerticalAnchor,
-        bottom: ConstraintLayoutBaseScope.HorizontalAnchor,
-        startMargin: Dp = 0.dp,
-        topMargin: Dp = 0.dp,
-        endMargin: Dp = 0.dp,
-        bottomMargin: Dp = 0.dp,
-        @FloatRange(from = 0.0, to = 1.0) horizontalBias: Float = 0.5f,
-        @FloatRange(from = 0.0, to = 1.0) verticalBias: Float = 0.5f
-    ) {
-        linkTo(start, end, startMargin, endMargin, horizontalBias)
-        linkTo(top, bottom, topMargin, bottomMargin, verticalBias)
-    }
-
-    /**
-     * Adds all start, top, end, bottom links towards the corresponding anchors of [other].
-     * This will center the current layout inside or around (depending on size) [other].
-     */
-    fun centerTo(other: ConstrainedLayoutReference) {
-        linkTo(other.start, other.top, other.end, other.bottom)
-    }
-
-    /**
-     * Adds start and end links towards the corresponding anchors of [other].
-     * This will center horizontally the current layout inside or around (depending on size)
-     * [other].
-     */
-    fun centerHorizontallyTo(
-        other: ConstrainedLayoutReference,
-        @FloatRange(from = 0.0f.toDouble(), to = 1.0f.toDouble()) bias: Float = 0.5f
-    ) {
-        linkTo(other.start, other.end, bias = bias)
-    }
-
-    /**
-     * Adds top and bottom links towards the corresponding anchors of [other].
-     * This will center vertically the current layout inside or around (depending on size)
-     * [other].
-     */
-    fun centerVerticallyTo(
-        other: ConstrainedLayoutReference,
-        @FloatRange(from = 0.0f.toDouble(), to = 1.0f.toDouble()) bias: Float = 0.5f
-    ) {
-        linkTo(other.top, other.bottom, bias = bias)
-    }
-
-    /**
-     * Adds start and end links towards a vertical [anchor].
-     * This will center the current layout around the vertical [anchor].
-     */
-    fun centerAround(anchor: ConstraintLayoutBaseScope.VerticalAnchor) {
-        linkTo(anchor, anchor)
-    }
-
-    /**
-     * Adds top and bottom links towards a horizontal [anchor].
-     * This will center the current layout around the horizontal [anchor].
-     */
-    fun centerAround(anchor: ConstraintLayoutBaseScope.HorizontalAnchor) {
-        linkTo(anchor, anchor)
-    }
-
-    /**
-     * Set a circular constraint relative to the center of [other].
-     * This will position the current widget at a relative angle and distance from [other].
-     */
-    fun circular(other: ConstrainedLayoutReference, angle: Float, distance: Dp) {
-        tasks.add { state ->
-            state.constraints(id)
-                .circularConstraint(other.id, angle, state.convertDimension(distance).toFloat())
-        }
-    }
-
-    /**
-     * Clear the constraints on the horizontal axis (left, right, start, end).
-     */
-    fun clearHorizontal() {
-        tasks.add { state ->
-            state.constraints(id).clearHorizontal()
-        }
-    }
-
-    /**
-     * Clear the constraints on the vertical axis (top, bottom, baseline).
-     */
-    fun clearVertical() {
-        tasks.add { state ->
-            state.constraints(id).clearVertical()
-        }
-    }
-
-    /**
-     * Clear all constraints (vertical, horizontal, circular).
-     */
-    fun clearConstraints() {
-        tasks.add { state ->
-            state.constraints(id).clear()
-        }
-    }
-
-    internal companion object {
-        val verticalAnchorFunctions:
-                Array<Array<ConstraintReference.(Any, LayoutDirection) -> ConstraintReference>> =
-            arrayOf(
-                arrayOf(
-                    { other, layoutDirection ->
-                        clearLeft(layoutDirection); leftToLeft(other)
-                    },
-                    { other, layoutDirection ->
-                        clearLeft(layoutDirection); leftToRight(other)
-                    }
-                ),
-                arrayOf(
-                    { other, layoutDirection ->
-                        clearRight(layoutDirection); rightToLeft(other)
-                    },
-                    { other, layoutDirection ->
-                        clearRight(layoutDirection); rightToRight(other)
-                    }
-                )
-            )
-
-        private fun ConstraintReference.clearLeft(layoutDirection: LayoutDirection) {
-            leftToLeft(null)
-            leftToRight(null)
-            when (layoutDirection) {
-                LayoutDirection.Ltr -> {
-                    startToStart(null); startToEnd(null)
-                }
-                LayoutDirection.Rtl -> {
-                    endToStart(null); endToEnd(null)
-                }
-            }
-        }
-
-        private fun ConstraintReference.clearRight(layoutDirection: LayoutDirection) {
-            rightToLeft(null)
-            rightToRight(null)
-            when (layoutDirection) {
-                LayoutDirection.Ltr -> {
-                    endToStart(null); endToEnd(null)
-                }
-                LayoutDirection.Rtl -> {
-                    startToStart(null); startToEnd(null)
-                }
-            }
-        }
-
-        /**
-         * Converts the index (-2 -> start, -1 -> end, 0 -> left, 1 -> right) to an index in
-         * the arrays above (0 -> left, 1 -> right).
-         */
-        // TODO(popam, b/157886946): this is temporary until we can use CL's own RTL handling
-        fun verticalAnchorIndexToFunctionIndex(index: Int, layoutDirection: LayoutDirection) =
-            when {
-                index >= 0 -> index // already left or right
-                layoutDirection == LayoutDirection.Ltr -> 2 + index // start -> left, end -> right
-                else -> -index - 1 // start -> right, end -> left
-            }
-
-        val horizontalAnchorFunctions:
-                Array<Array<ConstraintReference.(Any) -> ConstraintReference>> = arrayOf(
-            arrayOf(
-                { other -> topToBottom(null); baselineToBaseline(null); topToTop(other) },
-                { other -> topToTop(null); baselineToBaseline(null); topToBottom(other) }
-            ),
-            arrayOf(
-                { other -> bottomToBottom(null); baselineToBaseline(null); bottomToTop(other) },
-                { other -> bottomToTop(null); baselineToBaseline(null); bottomToBottom(other) }
-            )
-        )
-        val baselineAnchorFunction: ConstraintReference.(Any) -> ConstraintReference =
-            { other ->
-                topToTop(null)
-                topToBottom(null)
-                bottomToTop(null)
-                bottomToBottom(null)
-                baselineToBaseline(other)
-            }
-    }
-}
-
-interface VerticalAnchorable {
-    fun linkTo(anchor: ConstraintLayoutBaseScope.VerticalAnchor, margin: Dp = 0.dp)
-}
-
-interface HorizontalAnchorable {
-    fun linkTo(anchor: ConstraintLayoutBaseScope.HorizontalAnchor, margin: Dp = 0.dp)
-}
-
-interface BaselineAnchorable {
-    fun linkTo(anchor: ConstraintLayoutBaseScope.BaselineAnchor, margin: Dp = 0.dp)
-}
-
-@LayoutScopeMarker
-@Stable
-class HorizontalChainScope internal constructor(internal val id: Any) {
-    internal val tasks = mutableListOf<(State) -> Unit>()
-    internal fun applyTo(state: State) = tasks.forEach { it(state) }
-
-    /**
-     * Reference to the [ConstraintLayout] itself, which can be used to specify constraints
-     * between itself and its children.
-     */
-    val parent = ConstrainedLayoutReference(SolverState.PARENT)
-
-    /**
-     * The start anchor of the layout - can be constrained using [VerticalAnchorable.linkTo].
-     */
-    val start: VerticalAnchorable = HelperVerticalAnchorable(id, -2)
-
-    /**
-     * The left anchor of the layout - can be constrained using [VerticalAnchorable.linkTo].
-     */
-    val absoluteLeft: VerticalAnchorable = HelperVerticalAnchorable(id, 0)
-
-    /**
-     * The end anchor of the layout - can be constrained using [VerticalAnchorable.linkTo].
-     */
-    val end: VerticalAnchorable = HelperVerticalAnchorable(id, -1)
-
-    /**
-     * The right anchor of the layout - can be constrained using [VerticalAnchorable.linkTo].
-     */
-    val absoluteRight: VerticalAnchorable = HelperVerticalAnchorable(id, 1)
-
-    internal inner class HelperVerticalAnchorable internal constructor(
-        private val id: Any,
-        private val index: Int
-    ): VerticalAnchorable {
-        override fun linkTo(anchor: ConstraintLayoutBaseScope.VerticalAnchor, margin: Dp) {
-            tasks.add { state ->
-                with(state.helper(id, androidx.constraintlayout.core.state.State.Helper.HORIZONTAL_CHAIN)) {
-                    val layoutDirection = state.layoutDirection
-                    val index1 =
-                        ConstrainScope.verticalAnchorIndexToFunctionIndex(
-                            index,
-                            layoutDirection
-                        )
-                    val index2 =
-                        ConstrainScope.verticalAnchorIndexToFunctionIndex(
-                            anchor.index,
-                            layoutDirection
-                        )
-                    ConstrainScope.verticalAnchorFunctions[index1][index2]
-                        .invoke(this, anchor.id, state.layoutDirection)
-                        .margin(margin)
-                    apply()
-                }
-            }
-        }
     }
 }
