+++ constraintlayout/core/src/main/java/androidx/constraintlayout/core/state/helpers/HorizontalChainReference.java
@@ -43,8 +43,10 @@ public class HorizontalChainReference extends ChainReference {
                 } else if (mStartToEnd != null) {
                     first.startToEnd(mStartToEnd).margin(mMarginStart);
                 } else if (mLeftToLeft != null) {
+                    // TODO: Hack until we support RTL properly
                     first.startToStart(mLeftToLeft).margin(mMarginLeft);
                 } else if (mLeftToRight != null) {
+                    // TODO: Hack until we support RTL properly
                     first.startToEnd(mLeftToRight).margin(mMarginLeft);
                 } else {
                     first.startToStart(State.PARENT);
@@ -63,8 +65,10 @@ public class HorizontalChainReference extends ChainReference {
             } else if (mEndToEnd != null) {
                 previous.endToEnd(mEndToEnd).margin(mMarginEnd);
             } else if (mRightToLeft != null) {
+                // TODO: Hack until we support RTL properly
                 previous.endToStart(mRightToLeft).margin(mMarginRight);
             }else if (mRightToRight != null) {
+                // TODO: Hack until we support RTL properly
                 previous.endToEnd(mRightToRight).margin(mMarginRight);
             } else {
                 previous.endToEnd(State.PARENT);
