+++ projects/ComposeConstraintLayout/app/src/main/java/com/example/constraintlayout/verification/dsl/Chains.kt
@@ -93,21 +93,7 @@ fun Test11() {
 
 @Preview
 @Composable
-fun Test12() {
-    MyConstrainedChain()
-}
-
-@Preview
-@Composable
-fun Test13() {
-    CompositionLocalProvider(LocalLayoutDirection provides LayoutDirection.Rtl) {
-        MyConstrainedChain()
-    }
-}
-
-@Suppress("NOTHING_TO_INLINE")
-@Composable
-private inline fun MyConstrainedChain() {
+fun Test12() { // Constraint an horizontal chain
     val constraintSet = ConstraintSet {
         val box1 = createRefFor("box1")
         val box2 = createRefFor("box2")
@@ -134,6 +120,94 @@ private inline fun MyConstrainedChain() {
             start.linkTo(box3.end)
         }
     }
+    ThreeBoxChainLayout(constraintSet = constraintSet)
+}
+
+@Preview
+@Composable
+fun Test13() { // Constrain a vertical chain
+    val constraintSet = ConstraintSet {
+        val box1 = createRefFor("box1")
+        val box2 = createRefFor("box2")
+        val box3 = createRefFor("box3")
+        val chain1 = createVerticalChain(box1, box2, chainStyle = ChainStyle.Spread)
+
+        constrain(box1) {
+            width = Dimension.value(20.dp)
+            height = Dimension.value(20.dp)
+            centerHorizontallyTo(parent)
+        }
+        constrain(box2) {
+            width = Dimension.value(20.dp)
+            height = Dimension.value(20.dp)
+            centerHorizontallyTo(box1)
+        }
+        constrain(box3) {
+            width = Dimension.value(20.dp)
+            height = Dimension.value(200.dp)
+            top.linkTo(parent.top, 12.dp)
+            start.linkTo(parent.start, 12.dp)
+        }
+        constrain(chain1) {
+            top.linkTo(box3.bottom)
+        }
+    }
+    ThreeBoxChainLayout(constraintSet = constraintSet)
+}
+
+@Preview
+@Composable
+fun Test14() { // Constrain a chain with the Modifier DSL
+    Column {
+        ConstraintLayout(
+            modifier = Modifier
+                .fillMaxWidth()
+                .weight(1f, fill = true)
+        ) {
+            val (box1, box2, box3) = createRefs()
+            val chain1 = createHorizontalChain(box1, box2)
+            constrain(chain1) {
+                start.linkTo(box3.end)
+            }
+            Box(
+                modifier = Modifier
+                    .background(Color.Red)
+                    .constrainAs(box1) {
+                        width = Dimension.value(20.dp)
+                        height = Dimension.value(20.dp)
+                        centerVerticallyTo(parent)
+                    }
+            )
+            Box(
+                modifier = Modifier
+                    .background(Color.Blue)
+                    .constrainAs(box2) {
+                        width = Dimension.value(20.dp)
+                        height = Dimension.value(20.dp)
+                        centerVerticallyTo(box1)
+                    }
+            )
+            Box(
+                modifier = Modifier
+                    .background(Color.Green)
+                    .constrainAs(box3) {
+                        width = Dimension.value(200.dp)
+                        height = Dimension.value(20.dp)
+                        top.linkTo(parent.top, 12.dp)
+                        start.linkTo(parent.start, 12.dp)
+                    }
+            )
+
+        }
+        Button(onClick = { }) {
+            Text(text = "Run")
+        }
+    }
+}
+
+@Suppress("NOTHING_TO_INLINE")
+@Composable
+private inline fun ThreeBoxChainLayout(constraintSet: ConstraintSet) {
     Column {
         ConstraintLayout(
             modifier = Modifier
